+++ packages/frontend/editor-ui/src/components/canvas/experimental/experimentalNdv.utils.ts
@@ -1,5 +1,37 @@
-import type { INodeProperties } from 'n8n-workflow';
+import type { INodeUi } from '@/Interface';
+import type { I18nClass } from '@n8n/i18n';
+import type { INodeProperties, INodeTypeDescription } from 'n8n-workflow';
 
 export function shouldShowParameter(item: INodeProperties): boolean {
 	return item.name.match(/resource|authentication|operation/i) === null;
+}
+
+export function getNodeSubTitleText(
+	node: INodeUi,
+	nodeType: INodeTypeDescription,
+	includeOperation: boolean,
+	t: I18nClass,
+) {
+	if (node.disabled) {
+		return `(${t.baseText('node.disabled')})`;
+	}
+
+	const displayName = nodeType.displayName ?? '';
+
+	if (!includeOperation) {
+		return displayName;
+	}
+
+	const selectedOperation = node.parameters.operation;
+	const selectedOperationDisplayName =
+		selectedOperation &&
+		nodeType.properties
+			.find((p) => p.name === 'operation')
+			?.options?.find((o) => 'value' in o && o.value === selectedOperation)?.name;
+
+	if (!selectedOperationDisplayName) {
+		return displayName;
+	}
+
+	return `${displayName}: ${selectedOperationDisplayName}`;
 }
